Remove canUseRelativeOffsets
