+++ frontend/src/metabase/querying/filters/components/FilterModal/FilterModalBody/ColumnFilterGroup/ColumnFilterGroup.tsx
@@ -1,6 +1,6 @@
 import { useMemo, useState } from "react";
 
-import type { ColumnItem } from "metabase/querying/filters/types";
+import type { ColumnItem } from "metabase/querying/filters/hooks/use-filter-modal";
 import { DelayGroup } from "metabase/ui";
 import type * as Lib from "metabase-lib";
