+++ frontend/src/metabase/querying/filters/components/FilterModal/FilterModalBody/ColumnFilterItem/ColumnFilterItem.tsx
@@ -1,4 +1,4 @@
-import type { ColumnItem } from "metabase/querying/filters/types";
+import type { ColumnItem } from "metabase/querying/filters/hooks/use-filter-modal";
 import * as Lib from "metabase-lib";
 
 import { ColumnFilterSection } from "../../ColumnFilterSection";
