+++ frontend/src/metabase/querying/filters/components/FilterModal/FilterModalBody/SegmentFilterItem/SegmentFilterItem.tsx
@@ -1,4 +1,4 @@
-import type { SegmentItem } from "metabase/querying/filters/types";
+import type { SegmentItem } from "metabase/querying/filters/hooks/use-filter-modal";
 import type * as Lib from "metabase-lib";
 
 import { SegmentFilterEditor } from "../../SegmentFilterEditor";
