+++ frontend/src/metabase/querying/filters/hooks/use-boolean-option-filter/types.ts
@@ -5,6 +5,7 @@ export type OptionType = "true" | "false" | "is-null" | "not-null";
 
 export interface OperatorOption
   extends FilterOperatorOption<Lib.BooleanFilterOperator> {
+  name: string;
   type: OptionType;
   isAdvanced?: boolean;
 }
