--- frontend/src/metabase/querying/filters/hooks/use-filter-modal/utils/index.ts
@@ -1,1 +1,0 @@
-export * from "./search";
