+++ frontend/src/metabase/querying/filters/utils/column-groups.ts
@@ -1,0 +1,10 @@
+import type * as Lib from "metabase-lib";
+
+export function getGroupName(
+  groupInfo: Lib.ColumnGroupDisplayInfo,
+  stageIndex: number,
+) {
+  return groupInfo.isMainGroup && stageIndex > 1
+    ? `${groupInfo.displayName} (${stageIndex})`
+    : groupInfo.displayName;
+}
