--- frontend/src/metabase/querying/filters/utils/index.ts
@@ -1,2 +1,0 @@
-export * from "./filters";
-export * from "./operators";
